logging_service/requirements.txt
python-dateutil>=2.8.2
geopy>=2.2.0
flask-sqlalchemy>=3.1.0
kafka-python>=2.1.2
confluent_kafka>=2.8.2
elasticsearch>=8.17.2
psycopg2-binary>=2.9.0